Create README.md
We were instructed to works in teams for the ERD portion and to individually create our own SQL database. It was fluid in the beginning when exchanging ideas on how to map out the ERD. We ran into some confusion when needing to be more descriptive on whether a billing represented billing information versus billing info for example. I learned that being descriptive and specific with every function is extremely important in how each table connects and interacts. My create and insert database lists more customers coming into the dealership for car maintenance services since more complex categories contribute to the customer, mechanic, services needed, and labor.